Fix bug accessing the last variation value for the first day
Since it doesn't make sense to access the previous value of the first day on which the data recording began, the first row containing the "dash" value has been swapped with the last. This is due to the fact that the values are displayed in the order of the most recent datapoints.